pyproject.toml
[project]
name = "exp-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.3",
    "pillow>=11.2.1",
]
